Rename files and change code
-create folder fileAnalyzer

-Change code in Tests, add file.separator
