+++ src/test/java/com/maiboroda/fileAnalyzer/FileAnalyzerITest.java
@@ -1,6 +1,9 @@
-package com.maiboroda;
+package com.maiboroda.fileAnalyzer;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
+import java.io.File;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -8,23 +11,35 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 
-public class FileAnalyzerITest {
-    FileAnalyzer fileAnalyzer = new FileAnalyzer();
-    private final String path = "src/test/resources/text.txt";
+class FileAnalyzerITest {
+    FileAnalyzer fileAnalyzer;
+    private final String PATH = "src" + File.separator + "resources" + File.separator +  "text.txt";
+
+    @BeforeEach
+    void setUp(){
+        fileAnalyzer = new FileAnalyzer();
+    }
 
     @Test
     void testReadFile() {
+        String text = fileAnalyzer.readPath(PATH);
+
         assertFalse(text.isEmpty());
         assertTrue(text.toLowerCase().contains("java"));
     }
 
     @Test
     void testSplitToSentences() {
+        String text = fileAnalyzer.readPath(PATH);
+        List<String> sentences = fileAnalyzer.splitToSentences(text);
+
         assertEquals(10, sentences.size());
     }
 
     @Test
     void testSearchWordInSentences() {
+        String text = fileAnalyzer.readPath(PATH);
+        List<String> sentences = fileAnalyzer.splitToSentences(text);
         List<String> result = fileAnalyzer.searchWordInSentences(sentences, "java");
 
         assertFalse(result.isEmpty());
@@ -33,11 +48,10 @@ public class FileAnalyzerITest {
 
     @Test
     void testCountWord() {
+        String text = fileAnalyzer.readPath(PATH);
+        List<String> sentences = fileAnalyzer.splitToSentences(text);
         int count = fileAnalyzer.countWord(sentences, "java");
 
         assertEquals(10, count);
     }
-
-    private String text = fileAnalyzer.readPath(path);
-    private List<String> sentences = fileAnalyzer.splitToSentences(text);
 }
