+++ src/test/java/com/maiboroda/fileAnalyzer/FileAnalyzerTest.java
@@ -1,6 +1,7 @@
-package com.maiboroda;
+package com.maiboroda.fileAnalyzer;
 
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
@@ -13,9 +14,13 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class FileAnalyzerTest {
+class FileAnalyzerTest {
+    FileAnalyzer fileAnalyzer;
 
-    FileAnalyzer fileAnalyzer = new FileAnalyzer();
+    @BeforeEach
+    void setUp(){
+        fileAnalyzer = new FileAnalyzer();
+    }
 
     @Test
     void testValidateWordWithEmptyWord() {
